### 2022-12-05 #65 @Robaina
Hello, I would like to submit the package described above to pyOpenSci. Not sure if it fits your scope. Thanks!
