### 2022-12-06 #65 @NickleDave
Hi @Robaina and welcome to pyOpenSci!

Thank you for filling out the submission in a detailed thoughtful way.
My initial impression is that, yes, Pynteny is in scope for pyOpenSci.
I agree that it falls under the heading of the "data extraction" in the sense that Pynteny allows users to query a database with HMMs.

> I submitted this package for publication at JOSS a few days back.
We should coordinate with them. Our review processes are complementary. So far we've usually done the pyOS review first and then we pass off to JOSS. For example, see discussion on the pyGMT review here: https://github.com/pyOpenSci/software-review/issues/43 or the completed devicely review: https://github.com/pyOpenSci/software-review/issues/37

If that makes sense, please let me know, and please reference this issue on the open JOSS review issue

I will need to do more detailed editor checks but I wanted to welcome you and give you an initial answer as soon as possible. I will follow up tomorrow. Thank you @Robaina!
