### 2022-12-06 #65 @NickleDave
> I'm looking forward to being part of the pyOpenSci community :)

Glad to hear it 🙌

> I'll reference this issue in the open JOSS review issue.

I saw you commented there, thank you. I should have made it more clear, typically we review first so you do not need to go through two reviews. @lwasser asked me to clarify. I will comment on that review and link for the editor so they're aware.

On my to-do list to do more thorough checks later!
