### 2022-12-07 #65 @NickleDave
Hi again @Robaina I am familiarizing myself with the project a little more before I ask you to make a full submission; I want to make sure I understand how it works.

Good things:
- the README looks good overall, it has all the sections we would expect, and the badges and images are both helpful and attractive
- Your docs look great and so do the examples.

Some feedback:
 - We would want the docs to be a Sphinx site, not a wiki, as stated in our editor checks [here](https://www.pyopensci.org/peer-review-guide/software-peer-review-guide/editor-in-chief-guide.html#next-they-will-use-the-template-below-in-the-issue) although overall the docs are very readable. I would also recommend copying example snippets of using the cli so they are one of the very first thing a potential user sees on your README; they should know immediately what the interface actually looks like
- We would definitely want you to use a pyproject.toml that declares your dependencies as metadata, instead of the deprecated setup.py approach. Ideally with new-style metadata as described in the `flit` docs: https://flit.pypa.io/en/stable/pyproject_toml.html#new-style-metadata. For more on why we require a pyproject.toml, see [this post from Titus Brown](http://ivory.idyll.org/blog/2021-transition-to-pyproject.toml-example.html), and [this much more detailed post from Paul Anzel](https://blog.ganssle.io/articles/2021/10/setup-py-deprecated.html).
- Part of the purpose of pyOpenSci is to standardize project structure; there's no right answer but it is definitely uncommon to see a directory with the package name that then has a src/ directory inside of it. Usually you would either just have your source code in `pynteny` or you would have `src/pynteny`. See this post: https://hynek.me/articles/testing-packaging/. Similarly, while people do sometimes have tests along code, it's often preferable to completely isolate your tests so you know they are testing *installed* code. I would suggest switching to `src/pyntenty` or `pynteny` and moving your `tests` to the project root.
- Right now if I'm not mistaken it looks like you are limited to conda because that's the only way you can inspire HMMER? Is that correct? More about this below

Questions:
- Can you explain how `pynteny` compares to `PyHMMER`? Is there overlap in functionality there?
https://pyhmmer.readthedocs.io/en/stable/
Partly I ask because the developers of HMMER and PyHMMER are potential reviewers.
- I'm also asking about `PyHMMER` because it looks like you might be able to declare it as a dependency, so that your package would not be limited to only conda and Linux users; `PyHMMER` is pip installable so `Pynteny` would be pip installable as well. They appear to have some docs for developers here: https://pyhmmer.readthedocs.io/en/stable/examples/index.html
