### 2022-12-07 #65 @Robaina
Hi @NickleDave,

thank you for your careful first review of the project, here are my thoughts:

> * We would want the docs to be a Sphinx site, not a wiki, as stated in our editor checks [here](https://www.pyopensci.org/peer-review-guide/software-peer-review-guide/editor-in-chief-guide.html#next-they-will-use-the-template-below-in-the-issue) although overall the docs are very readable. I would also recommend copying example snippets of using the cli so they are one of the very first things a potential user sees on your README; they should know immediately what the interface actually looks like

That makes sense. Will port the wiki to a Sphinx site. I agree on including cli snippets!

> * We would definitely want you to use a pyproject.toml that declares your dependencies as metadata, instead of the deprecated setup.py approach. Ideally with new-style metadata as described in the `flit` docs: https://flit.pypa.io/en/stable/pyproject_toml.html#new-style-metadata. For more on why we require a pyproject.toml, see [this post from Titus Brown](http://ivory.idyll.org/blog/2021-transition-to-pyproject.toml-example.html), and [this much more detailed post from Paul Anzel](https://blog.ganssle.io/articles/2021/10/setup-py-deprecated.html).

Will do. A good opportunity to finally get rid of the deprecated setup.py.

> * Part of the purpose of pyOpenSci is to standardize project structure; there's no right answer but it is definitely uncommon to see a directory with the package name that then has a src/ directory inside of it. Usually you would either just have your source code in `pynteny` or you would have `src/pynteny`. See this post: https://hynek.me/articles/testing-packaging/. Similarly, while people do sometimes have tests along code, it's often preferable to completely isolate your tests so you know they are testing _installed_ code. I would suggest switching to `src/pyntenty` or `pynteny` and moving your `tests` to the project root.

Yeah, actually I wasn't following any standard structure in the project.  Seems reasonable to change it to what you suggest.

> * Right now if I'm not mistaken it looks like you are limited to conda because that's the only way you can inspire HMMER? Is that correct? More about this below

It started like that, yes. Pynteny also depends on Prodigal, which is available in [bioconda](https://anaconda.org/bioconda/prodigal). I made wrappers for both to be run within Pynteny. Although providing a wrapper was more of a necessity rather than a designed functionality back then.

However, I have checked and there is a python binding to prodigal, which is pip installable: [Pyrodigal](https://pypi.org/project/pyrodigal/). All other dependencies are also both pip and conda installable, so I could make Pynteny also pip installable by switching from using HMMER and Prodigal to using PyHMMER and Pyrodigal. I would have to delve into it since I believe that PyHMMER and HMMER output formats are structured differently.

I plan to submit Pynteny to the [bioconda channel](https://bioconda.github.io/) since it is dedicated to software within the biological sciences. Bioconda would make Pynteny available for MacOS as well as Linux, but not for Windows (that is, outside using WSL). However, I think that it is worth trying PyHMMER and Pyrodigal so I can make Pynteny also pip installable and potentially usable also in Windows without WSL

>
> Questions:
>
> * Can you explain how `pynteny` compares to `PyHMMER`? Is there overlap in functionality there?
>   https://pyhmmer.readthedocs.io/en/stable/
>   Partly I ask because the developers of HMMER and PyHMMER are potential reviewers.

Currently, Pynteny provides a simple CLI wrapper to HMMER (contrary to the sophisticated interface provided by PyHMMER). This is not the main functionality of Pynteny (and could be entirely removed by using PyHMMER instead of HMMER).

Pynteny uses HMMER (potentially PyHMMER) as a tool to run sequence profile HMM searches, which is a step required to then filter those searches by the syntenic structure provided by the user.  So, to clarify, these would be the main functionalities provided by Pynteny:

1) Construct a peptide database from DNA assembly data in which peptide labels contain positional information (gene number within contig / chromosome, and contig name). DNA translation to peptide sequences is carried out by Prodigal / Pyrodigal.

2) Search the peptide database by a predefined syntenic structure (i.e., gene spatial arrangement) using profile HMMs. Here, HMMER / PyHMMER is employed to search the peptide database for every gene/HMM in the provided syntenic structure. Pynteny  would then take into account the positional information and the syntenic structure (including additional optional constraints like DNA strand specificity and alternative HMMs for the  same gene) to filter HMMER results and retain only those sequence hits that are aligned with the provided syntenic structure.

3) Additionally, Pynteny facilitates downloading a large profile HMM database ([PGAP](https://github.com/ncbi/pgap)) as well as "translating" a syntenic structure defined with gene symbols to another one containing matching profile HMMs from the PGAP database.

> * I'm also asking about `PyHMMER` because it looks like you might be able to declare it as a dependency, so that your package would not be limited to only conda and Linux users; `PyHMMER` is pip installable so `Pynteny` would be pip installable as well. They appear to have some docs for developers here: https://pyhmmer.readthedocs.io/en/stable/examples/index.html

See my answers above.

---
Hope my comments help clarify Pynteny's functionality. Please, don't hesitate to ask any further questions.
Thanks!
