### 2022-12-07 #65 @lwasser
@Robaina so so psyched to have you here. Just a note that if you need help with any of the above items we are happy to support you. you [can post questions here in this issue OR on discourse if you wish](https://pyopensci.discourse.group/c/coding-help/10). I just went through updating a package with docs and pyproject.toml and found that the online documentation was a bit confusing so just reach out if you have questions!

in the future we will have better documentation to support making changes like the ones discussed above!
