### 2022-12-07 #65 @Robaina
Hi @lwasser,

thanks! You and @NickleDave have been very supportive. Alright, I'll try discourse as it may help future developers.
