### 2022-12-07 #65 @Robaina
Hi @NickleDave,

I have thought about relocating the tests directory to root as you suggested. While I see the advantages you listed by locating tests there,   I'm not entirely convinced that this is the best option in this case. This is because currently tests are installed (as a subpackage) alongside pynteny and can be run with the subcommand `Pynteny tests`, which I think it's a nice functionality for users to check that installation went fine. I am not sure I want to get rid of this functionality to favor locating tests outside of the installation. Does this make sense to you?
