### 2022-12-08 #65 @NickleDave
Hi @Robaina thank you so much for your clear answers and for being open to feedback.

Your response makes me more confident Pynteny is in scope, since it also includes data retrieval, and my understand that it leverages functionality from both HMMER and Prodigal to provide a unique ability to query a peptide database with HMMs. Aside: do you have a sort of workflow diagram in your docs that captures this description, including which dependency does what? Might be helpful

Before we make a final decision, I just want to get a little feedback from devs in bioinformatics and genomics--I have asked in our Slack--but I am 90% sure we will ask you to go ahead with a submission. (I should have asked them before, I am sorry for not doing so in the first place.)

In the meantime, could I ask you to track any changes you are making as a result of this presubmission as issues, and then reference those issues on any submission? @lwasser may modify our survey as well to make sure we capture that (we are still figuring out process as we come back on line).

So far we have (correct me if I'm wrong):
- [ ] switch docs to Sphinx (I highly __highly__ suggesting using [myst](https://myst-parser.readthedocs.io/en/latest/), by the way, if you weren't already)
- [ ] restructure project: make docs/, src/ and tests/ folders (more on this below, I did see your question)
- [ ] test whether you can depend on PyHMMER and Pyrodigal to enable a pip-only build (that probably makes it easier and more reliable in the long run to generate packages for bioconda and conda-forge e.g. with [grayskull](https://github.com/conda-incubator/grayskull#introduction))
