### 2022-12-08 #65 @NickleDave
re: keeping tests in docs

This does not affect our decision on a submission; we don't have any hard and fast rules right now about test. As a developer I totally understand why you want to keep them there. The reasons I would suggest moving them out are
(1) you want to ideally isolate tests from source and run on an installed build (even if you don't do so during day-to-day development),
(2) it's great that you care about your users but most of them won't be developers and won't care about running tests, and the ones that do care will know how to clone your repo and run tests the usual way
(3) I'm not sure what running tests gives users unless you have a lot of edge cases that you need them to report to you about, and even in that case it might be better to add a helper function that dumps out env info to the terminal that they can paste into issues with an error message. I did look at your tests but nothing jumps out at me that makes me think you are often needing it to be in with the code

Most recent updated guides I know suggest putting tests in a separate directory:
https://docs.python-guide.org/writing/structure/
https://py-pkgs.org/05-testing
https://packaging.python.org/en/latest/tutorials/packaging-projects/#a-simple-project
https://cjolowicz.github.io/posts/hypermodern-python-02-testing/
https://doc.pytest.org/en/latest/explanation/goodpractices.html#tests-outside-application-code
https://github.com/pypa/sampleproject

Does that help?
