### 2022-12-08 #65 @Robaina
Hi @NickleDave ,

> Your response makes me more confident Pynteny is in scope, since it also includes data retrieval, and my understand that it leverages functionality from both HMMER and Prodigal to provide a unique ability to query a peptide database with HMMs. Aside: do you have a sort of workflow diagram in your docs that captures this description, including which dependency does what? Might be helpful

Excellent! Regarding the diagram:  I could do a simple one and include it in the index page of the docs.

> Before we make a final decision, I just want to get a little feedback from devs in bioinformatics and genomics--I have asked in our Slack--but I am 90% sure we will ask you to go ahead with a submission. (I should have asked them before, I am sorry for not doing so in the first place.)

Alright, no problem. I'll wait for the final decision.

> In the meantime, could I ask you to track any changes you are making as a result of this presubmission as issues, and then reference those issues on any submission? @lwasser may modify our survey as well to make sure we capture that (we are still figuring out process as we come back on line).

I already made some changes, but I can do that from now on.

> So far we have (correct me if I'm wrong):
>
> * [ ]  switch docs to Sphinx (I highly **highly** suggesting using [myst](https://myst-parser.readthedocs.io/en/latest/), by the way, if you weren't already)

Yeah, so I'll try myst as per your suggestion. Ideally, I want to use the same markdown files I already made but I think that shouldn't be a problem.

> * [ ]  restructure project: make docs/, src/ and tests/ folders (more on this below, I did see your question)

This is already been taken care of (not in the repo yet, though). Thanks for your comments about tests!

> * [ ]  test whether you can depend on PyHMMER and Pyrodigal to enable a pip-only build (that probably makes it easier and more reliable in the long run to generate packages for bioconda and conda-forge e.g. with [grayskull](https://github.com/conda-incubator/grayskull#introduction))

I have this in mind and will definitely test those packages. However, I do have a question: would making Pynteny pip-installable be a requisite for consideration at pySciOpen? Or rather a desirable change that I could do later on?

Thank you for all your suggestions. They have definitely  been very helpful
