+++ pynteny/__init__.py
@@ -1,1 +1,2 @@
 from pynteny.api import *
+from pynteny.cli import main
