+++ MANIFEST.in
@@ -1,6 +1,2 @@
-include pynteny/tests/*
-include pynteny/tests/test_data/*
-include pynteny/tests/test_data/hmms/*
-include pynteny/tests/test_data/test_assembly/*
-include pynteny/src/app/assets/*
-include pynteny/src/app/assets/img/*
+include pynteny/app/assets/*
+include pynteny/app/assets/img/*
