+++ .readthedocs.yaml
@@ -1,0 +1,28 @@
+# This is the main configuration file for Read the Docs.
+# For a full list of options, see:
+# https://docs.readthedocs.io/en/stable/config-file/v2.html
+
+# The version of Python used to build the documentation
+python:
+  version: 3.8
+
+# Where to find the source files
+source:
+  # The path to the root of the documentation source files
+  # This can be a local directory or a remote Git repository
+  path: /docs
+  
+  # The URL of the remote Git repository
+  # This is required if you are using a remote repository
+  # For example, for a GitHub repository:
+  # https://github.com/your-username/your-repository.git
+  git: https://github.com/Robaina/Pynteny.git
+
+# The version of the documentation to build
+# This can be a branch, tag, or commit hash
+version: latest
+
+# The format of the documentation files
+# Read the Docs supports a number of different file formats, including Markdown
+formats:
+  - markdown
