+++ ms/paper.bib
@@ -1,194 +1,0 @@
-@article{hmmer,
-    author = {Mistry, Jaina and Finn, Robert D. and Eddy, Sean R. and Bateman, Alex and Punta, Marco},
-    title = "{Challenges in homology search: HMMER3 and convergent evolution of coiled-coil regions}",
-    journal = {Nucleic Acids Research},
-    volume = {41},
-    number = {12},
-    pages = {e121-e121},
-    year = {2013},
-    month = {04},
-    abstract = "{Detection of protein homology via sequence similarity has important applications in biology, from protein structure and function prediction to reconstruction of phylogenies. Although current methods for aligning protein sequences are powerful, challenges remain, including problems with homologous overextension of alignments and with regions under convergent evolution. Here, we test the ability of the profile hidden Markov model method HMMER3 to correctly assign homologous sequences to \\&gt;13 000 manually curated families from the Pfam database. We identify problem families using protein regions that match two or more Pfam families not currently annotated as related in Pfam. We find that HMMER3 E-value estimates seem to be less accurate for families that feature periodic patterns of compositional bias, such as the ones typically observed in coiled-coils. These results support the continued use of manually curated inclusion thresholds in the Pfam database, especially on the subset of families that have been identified as problematic in experiments such as these. They also highlight the need for developing new methods that can correct for this particular type of compositional bias.}",
-    issn = {0305-1048},
-    doi = {10.1093/nar/gkt263},
-    url = {https://doi.org/10.1093/nar/gkt263}
-}
-
-@article{blast,
-	title = {Magic-{BLAST}, an accurate {RNA}-seq aligner for long and short reads},
-	volume = {20},
-	issn = {1471-2105},
-	url = {https://doi.org/10.1186/s12859-019-2996-x},
-	doi = {10.1186/s12859-019-2996-x},
-	abstract = {Next-generation sequencing technologies can produce tens of millions of reads, often paired-end, from transcripts or genomes. But few programs can align RNA on the genome and accurately discover introns, especially with long reads. We introduce Magic-BLAST, a new aligner based on ideas from the Magic pipeline.},
-	number = {1},
-	urldate = {2022-11-03},
-	journal = {BMC Bioinformatics},
-	author = {Boratyn, Grzegorz M. and Thierry-Mieg, Jean and Thierry-Mieg, Danielle and Busby, Ben and Madden, Thomas L.},
-	month = jul,
-	year = {2019},
-	keywords = {Alignment, BLAST, RNA-seq},
-	pages = {405}
-}
-
-@article{hmms-a,
-	title = {Hidden {Markov} model speed heuristic and iterative {HMM} search procedure},
-	volume = {11},
-	issn = {1471-2105},
-	url = {https://doi.org/10.1186/1471-2105-11-431},
-	doi = {10.1186/1471-2105-11-431},
-	abstract = {Profile hidden Markov models (profile-HMMs) are sensitive tools for remote protein homology detection, but the main scoring algorithms, Viterbi or Forward, require considerable time to search large sequence databases.},
-	number = {1},
-	urldate = {2022-11-03},
-	journal = {BMC Bioinformatics},
-	author = {Johnson, L. Steven and Eddy, Sean R. and Portugaly, Elon},
-	month = aug,
-	year = {2010},
-	keywords = {Entropy Weighting, Iterative Search, Profile Hide Markov Model, Search Time, Test Database},
-	pages = {431}
-}
-
-@article{hmms-b,
-	title = {What is a hidden {Markov} model?},
-	volume = {22},
-	copyright = {2004 Nature Publishing Group},
-	issn = {1546-1696},
-	url = {https://www.nature.com/articles/nbt1004-1315},
-	doi = {10.1038/nbt1004-1315},
[…]
