+++ pynteny/__init__.py
@@ -1,2 +1,0 @@
-from pynteny.api import *
-from pynteny.cli import main

[…]
