+++ docs2/app.md
@@ -1,0 +1,15 @@
+# Description
+
+Run Pynteny as a local app in the web browser.
+
+# Usage:
+
+```bash
+usage: pynteny app [-h] 
+
+```
+# Arguments
+
+|short|long|default|help|
+| :--- | :--- | :--- | :--- |
+|`-h`|`--help`||show this help message and exit|
