+++ mkdocs.yml
@@ -5,3 +5,12 @@ site_author: Semidán Robaina
 nav:
     - Home: index.md
 theme: readthedocs
+# Add MathJax support:
+markdown_extensions:
+  - pymdownx.arithmatex:
+      generic: true
+
+extra_javascript:
+  - javascripts/mathjax.js
+  - https://polyfill.io/v3/polyfill.min.js?features=es6
+  - https://cdn.jsdelivr.net/npm/mathjax@3/es5/tex-mml-chtml.js
