+++ .github/workflows/documentation.yml
@@ -25,7 +25,7 @@ jobs:
         with:
           fetch-depth: 0
       - uses: actions/setup-python@v2
-      - run: pip install --upgrade pip && pip install mkdocs mkdocs-gen-files
+      - run: pip install --upgrade pip && pip install mkdocs mkdocs-gen-files pymdown-extensions
       - run: git config user.name 'github-actions[bot]' && git config user.email 'github-actions[bot]@users.noreply.github.com'
       - name: Publish docs
         run: mkdocs gh-deploy
