+++ docs/index.md
@@ -4,30 +4,29 @@ Pynteny is a Python package to query sequence databases for synteny blocks throu
 
 These are the available subcommands, run as ```pynteny <subcommand> <options>```:
 
-- [search](https://github.com/Robaina/Pynteny/wiki/search)
-- [build](https://github.com/Robaina/Pynteny/wiki/build)
-- [download](https://github.com/Robaina/Pynteny/wiki/download)
-- [parse](https://github.com/Robaina/Pynteny/wiki/parse)
-- [tests](https://github.com/Robaina/Pynteny/wiki/tests)
-- [app](https://github.com/Robaina/Pynteny/wiki/app)
+- [search](subcommands/search.md)
+- [build](subcommands/build.md)
+- [download](subcommands/download.md)
+- [parse](subcommands/parse.md)
+- [app](subcommands/app.md)
 
 ## General usage
 
-Pynteny's main subcommand, [`pynteny search`](https://github.com/Robaina/Pynteny/wiki/search), requires a peptide (ORF) sequence database in fasta format in which record labels contain positional information of each sequence with respect to their contig of origin. Additionally, labels must follow the following format:
+Pynteny's main subcommand, [`pynteny search`](subcommands/search.md), requires a peptide (ORF) sequence database in fasta format in which record labels contain positional information of each sequence with respect to their contig of origin. Additionally, labels must follow the following format:
 
 ```
 <genome ID>_<contig ID>_<gene position>_<locus start>_<locus end>_<strand>
 ```
 
-To facilitate the construction of this database, Pynteny provides the subcommand [`pynteny build`](https://github.com/Robaina/Pynteny/wiki/build), which takes as input a fasta file containing assembled nucleotide sequence data (or a single or a collection of genomes), such as that retrieved from MAG reconstruction pipelines.
+To facilitate the construction of this database, Pynteny provides the subcommand [`pynteny build`](subcommands/build.md), which takes as input a fasta file containing assembled nucleotide sequence data (or a single or a collection of genomes), such as that retrieved from MAG reconstruction pipelines.
 
-Additionally, [`pynteny search`](https://github.com/Robaina/Pynteny/wiki/search) requires either the set of profile Hidden Markov Models (HMMs) used in the provided synteny structure or a database of profile HMMs from which to retrieve the necessary HMMs. The user may provide their own set of HMMs. However, the user may also download the entire [PGAP HMM database](https://academic.oup.com/nar/article/49/D1/D1020/6018440) through [`pynteny download`](https://github.com/Robaina/Pynteny/wiki/download), which will take care of downloading and storing the download path for future usage (by default any time no additional HMMs are provided as an argument in [`pynteny search`](https://github.com/Robaina/Pynteny/wiki/search)).
+Additionally, [`pynteny search`](subcommands/search.md) requires either the set of profile Hidden Markov Models (HMMs) used in the provided synteny structure or a database of profile HMMs from which to retrieve the necessary HMMs. The user may provide their own set of HMMs. However, the user may also download the entire [PGAP HMM database](https://academic.oup.com/nar/article/49/D1/D1020/6018440) through [`pynteny download`](subcommands/download.md), which will take care of downloading and storing the download path for future usage (by default any time no additional HMMs are provided as an argument in [`pynteny search`](subcommands/search.md)).
 
 Once both the peptide database and the required HMMs are ready, the user can query the peptide database with a text string encoding the query synteny structure such as the following:
 
-$>HMM_a \\:\\: n_{ab} \\:\\: > (HMM_{b1} | HMM_{b2}) \\:\\: n_{bc} \\:\\: < HMM_c.$
[…]
