+++ docs/tests.md
@@ -1,15 +1,0 @@
-# Description
-
-Run Pynteny unit and integration tests.
-
-# Usage:
-
-```bash
-usage: pynteny tests [-h] 
-
-```
-# Arguments
-
-|short|long|default|help|
-| :--- | :--- | :--- | :--- |
-|`-h`|`--help`||show this help message and exit|
