+++ mkdocs.yml
@@ -4,6 +4,13 @@ repo_url: https://github.com/Robaina/Pynteny
 site_author: Semidán Robaina
 nav:
     - Home: index.md
+    - Subcommands:
+        - Search: search.md
+        - Build: build.md
+        - Parse: parse.md
+        - Download: download.md
+        - App: app.md
+    - 
 theme: readthedocs
 # Add MathJax support:
 markdown_extensions:
