### 2022-12-09 #65 @NickleDave
Happy to hear the suggestions are helpful!
I need to do a better job of not imposing my own opinions but I do really appreciate that you are open to feedback.

Just so we're on the same page: I am still expecting we will go forward with this as a submission, I am just waiting to hear back from subject matter experts. Thanks for your patience up until now and please expect that final decision early next week.

> would making Pynteny pip-installable be a requisite for consideration at pySciOpen?

Short answer: we want to be understanding of the current reality of Python packaging and we want to favor providing people access to good tools. So I would not say it's a requirement. Longer answer: we don't have official language on this in our guides right now and this is part of the revisions we're doing right now.

So effectively, no it's not a requirement.

I see that you did raise an issue about it. @lwasser noticed that Pyteny is on PyPI. I was able to pip install on Linux and at least run the download command from the cli. On mac with Python 3.7 I get `ImportError: cannot import name 'metadata' from 'importlib'`.

Given that plus our discussion above I think this means it seems very possible you can make Pynteny pip installable once you figure out any data format issues. Again we would not require you get this done before submission or somehow reject after review because you weren't able to get it done.

Does that make sense? Great to see all the progress you've already made just during this presubmission--new docs look good! Feel free to ask if you have further questions. Otherwise I will be back in touch early next week.
