### 2022-12-09 #65 @Robaina
Hi again :)

>I need to do a better job of not imposing my own opinions but I do really appreciate that you are open to feedback.

no problem at all, I did not read it as an "imposition". I asked that because I'm not entirely sure I can easily integrate PyHMMER /  Pyrodigal into Pynteny without major changes to the code. Also, using those packages means abandoning usage of the original implementations which I guess are more tested. Anyway, I haven't tested it yet, I just wanted to get an idea about your opinions on the topic.

> I see that you did raise an issue about it. @lwasser noticed that Pyteny is on PyPI. I was able to pip install on Linux and at least run the download command from the cli. On mac with Python 3.7 I get ImportError: cannot import name 'metadata' from 'importlib'.

Right. Sorry, Pynteny on pip is not functional. I thought I had deleted Pynteny from pip. I already did. This was an early attempt (hence the old setup.py file as well) but I abandoned it after realizing HMMER and prodigal were only on bioconda, so it made little sense to me to first install a conda environment with dependencies and then Pynteny via pip. Thus I continued with conda build and made a recipe.

If PyHMMER/Pyrodingal works well, I'll build a proper pip package  (using myproject.toml) and upload it to pip.

>Does that make sense? Great to see all the progress you've already made just during this presubmission--new docs look good! Feel free to ask if you have further questions. Otherwise I will be back in touch early next week.

It does. Again, thanks for the suggestions. I'm not a software engineer / developer by training so these suggestions are very useful.

Till next week then!
