### 2022-12-09 e6d347d
fixed path to examples
