+++ docs/index.md
@@ -30,7 +30,7 @@ Where $HMM_a$ represents the name of the HMM to be used (corresponding to the fi
 
 ## Examples
 
-There are some Jupyter notebooks with examples of how to use Pynteny's command-line interface as well as its Python API. Check them out in the [documentation](examples) or in the [GitHub repo](https://github.com/Robaina/Pynteny/tree/master/docs/examples).
+There are some Jupyter notebooks with examples of how to use Pynteny's command-line interface as well as its Python API. Check them out here in the examples section or in the [GitHub repo](https://github.com/Robaina/Pynteny/tree/master/docs/examples).
 ## Citation
 
 If you use this software, please cite it as below:
