### 2022-12-13 ac1dd88
filter.py docstrings
