+++ pynteny/filter.py
@@ -21,17 +21,18 @@ logger = logging.getLogger(__name__)
 
 
 class SyntenyPatternFilters():
+    """Methods to filter hmm hits in the  same contig by synteny structure
+       or collinearity. These filters are inputs to pandas.Dataframe.rolling method.
+    """
     def __init__(self, synteny_structure: str, unordered: bool = False) -> None:
-        """
-        Methods to filter hmm hits in the  same contig by synteny structure
-        or collinearity. To be applied in pandas.Series.rolling.apply method.
+        """Initialize filter class from synteny structure.
 
-        @param: synteny_structure: string coontaining the synteny structure
-                to be searched for
-        @param: unordered, boolean, whether the HMMs should be arranged in the
+        Args:
+            synteny_structure (str): synteny structure following style described in the docs.
+            unordered (bool, optional): whether the HMMs should be arranged in the
                 exact same order displayed in the synteny_structure or in 
                 any order If ordered, the filters would filter collinear rather
-                than syntenic structures.
+                than syntenic structures. Defaults to False.
         """
         parsed_structure = SyntenyParser.parseSyntenyStructure(synteny_structure)
         hmm_order_dict = dict(
@@ -58,12 +59,28 @@ class SyntenyPatternFilters():
         self._unordered = unordered
 
     def contains_hmm_pattern(self, data: pd.Series) -> int:
+        """Check if series items contain a profile HMM
+
+        Args:
+            data (pd.Series): a series resulting from calling rolling on a pandas column.
+
+        Returns:
+            int: 1 for True 0 for False.
+        """
         if self._unordered:
             return 1 if set(data.values) == set(self.hmm_code_order_pattern) else 0
         else:
             return 1 if data.values.tolist() == self.hmm_code_order_pattern else 0
 
     def contains_distance_pattern(self, data: pd.Series) -> int:
+        """_summary_
+
+        Args:
+            data (pd.Series): _description_
+
+        Returns:
+            int: _description_
+        """
         return 1 if all(
             [
                 (data_dist <= max_dist) and (data_dist > min_dist)
@@ -72,6 +89,14 @@ class SyntenyPatternFilters():
             ) else 0
 
     def contains_strand_pattern(self, data: pd.Series) -> int:
+        """_summary_
+
+        Args:
+            data (pd.Series): _description_
+
+        Returns:
+            int: _description_
+        """
         strand_comparisons = []
         for data_strand, pattern_strand in zip(data.values, self.strand_order_pattern):
             if pattern_strand != 0:
@@ -82,30 +107,28 @@ class SyntenyPatternFilters():
 
 
 class SyntenyHMMfilter():
-    """
-    Tools to search for synteny structures among sets of hmm models
+    """Tools to search for synteny structures among sets of hmm models
     """
     def __init__(self, hmm_hits: dict, synteny_structure: str, unordered: bool = True) -> None:
-        """
-        Search for contigs that satisfy the given gene synteny structure
+        """Search for contigs that satisfy the given gene synteny structure.
 
-        @param: hmm_hits, a dict of pandas DataFrames, as output by
+        Args:
+            hmm_hits (dict): a dict of pandas DataFrames, as output by
                 parseHMMsearchOutput with keys corresponding to hmm names
-
-        @param: synteny_structure, a str describing the desired synteny structure,
+            synteny_structure (str): a str describing the desired synteny structure,
                 structured as follows:
 
-                'hmm_a N_ab hmm_b'
+                '>hmm_a N_ab hmm_b bc <hmm_c'
 
                 where N_ab corresponds to the maximum number of genes separating 
[…]
