### 2022-12-13 086174e
hmm.py docstring
