+++ pynteny/utils.py
@@ -21,13 +21,14 @@ logger = logging.getLogger(__name__)
 
 
 class CommandArgs():
+    """Base class to hold command line arguments.
+    """
     def __init__(self, **kwargs):
         self.__dict__.update(kwargs)
 
         
 class ConfigParser():
-    """
-    Base configuration class
+    """Handle Pynteny configuration file.
     """
     def __init__(self, config_file: Path) -> None:
         self._config_file = Path(config_file)
@@ -35,15 +36,16 @@ class ConfigParser():
     
     @classmethod
     def get_default_config(cls):
-        """
-        Initialize ConfigParser with default config file
+        """Initialize ConfigParser with default config file.
         """
         return cls(cls.initialize_config_file())
     
     @staticmethod
     def initialize_config_file() -> Path:
-        """
-        initialize empty config file
+        """Initialize empty config file.
+
+        Returns:
+            Path: path to generated config file.
         """
         config_file = Path(Path(Path(__file__).parent).parent) / "config.json"
         if not config_file.exists():
@@ -60,36 +62,44 @@ class ConfigParser():
         return config_file
     
     def get_config_path(self) -> Path:
-        """
-        Show config file path
+        """Show config file path.
         """
         return self._config_file
 
     def get_config(self) -> dict:
-        """
-        load config file
+        """Load config file.
+
+        Returns:
+            dict: dict containing fields and values of config file.
         """
         with open(self._config_file, 'r') as file:
             config = json.loads(file.read())
         return config
 
     def write_config(self) -> None:
-        """
-        write config file
+        """Write config dict to file.
         """
         with open(self._config_file, 'w') as f:
             json.dump(self._config, f, indent=4)
 
     def update_config(self, key: str, value: str) -> None:
-        """
-        update config file
+        """Update config file
+
+        Args:
+            key (str): config file key name to be updated.
+            value (str): new value.
         """
         self._config[key] = value
         self.write_config()
 
     def get_field(self, key: str) -> str:
-        """
-        get field from config file
+        """Get field from config file.
+
+        Args:
+            key (str): key name to get the value from.
+
+        Returns:
+            str: key value.
         """
         return self._config[key]
 
@@ -98,9 +108,20 @@ def setDefaultOutputPath(input_path: Path, tag: str = None,
                          extension: str = None,
                          only_filename: bool = False,
                          only_basename: bool = False,
-                         only_dirname: bool = False) -> str:
-    """
-    Get default path to output file or directory
+                         only_dirname: bool = False) -> Path:
+    """Utility function to generate a default path to output file
+    or directory based on an input file name and path.
+
+    Args:
+        input_path (Path): path to input file.
+        tag (str, optional): text tag to be added to file name. Defaults to None.
+        extension (str, optional): change input file extension with this one. Defaults to None.
+        only_filename (bool, optional): output only default filename. Defaults to False.
+        only_basename (bool, optional): output only default basename (no extension). Defaults to False.
+        only_dirname (bool, optional): output only path to default output directory. Defaults to False.
+
+    Returns:
+        Path: a path or name to a default output file.
     """
     dirname = input_path.parent
     fname, ext = input_path.stem, input_path.suffix
@@ -122,8 +143,16 @@ def terminalExecute(command_str: str,
                     suppress_shell_output=False,
                     work_dir: Path = None,
                     return_output=False) -> subprocess.STDOUT:
-    """
[…]
