+++ docs/references.md
@@ -1,0 +1,1 @@
+:::pynteny.api
