+++ docs/references.md
@@ -1,1 +1,1 @@
-:::pynteny.api
+:::pynteny.filter
