+++ docs/references.md
@@ -1,1 +1,1 @@
-:::pynteny.filter
+::: pynteny.filter
