### 2022-12-14 #65 @NickleDave
Hi @Robaina!
My apologies for not getting back to you sooner, a bit of a crunch at work right now.
I also appreciate your patience overall--I should have given you a timeline sooner and then done checks all at once before getting back to you.

And of course we appreciate your enthusiasm!
We want this to be a good experience for you and we can tell you are highly motivated.

I did try to install as in the README last night and had some issues.
I'd like to suggest we proceed as follows:
- [ ] I will raise issues on your repo describing what I ran into when I tried to install following the README
- [ ] I would suggest you ask in our Discourse for help with resolving that. We do have some people that are more familiar with conda and we are happy to make sure they see your post
- [ ] It will probably help them/us/you if you can add files to the Pynteny repo that you use to publish your conda package--I'm guessing you're using conda-build with a meta.yaml but I don't see it in the repo?

Does that make sense? We do feel `Pynteny` is in scope and we want to go ahead with a submission but we definitely want to be sure we can install first

I have some meetings ~12-5 today but will raise those issues about what I ran into with installs by the end of the day at the latest
