+++ docs/references/api.md
@@ -1,0 +1,26 @@
+## API module
+::: pynteny.api
+
+## Module filter
+::: pynteny.filter
+
+## Module preprocessing
+::: pynteny.preprocessing
+
+## Module hmm
+::: pynteny.hmm
+
+## Module parser
+::: pynteny.parser
+
+## Module cli
+::: pynteny.cli
+
+## Module subcommands
+::: pynteny.subcommands
+
+## Module utils
+::: pynteny.utils
+
+## Module wrappers
+::: pynteny.wrappers
