+++ mkdocs.yml
@@ -14,7 +14,8 @@ nav:
          - Example CLI: examples/example_cli.ipynb
          - Example API: examples/example_api.ipynb
          - Sus operon: examples/example_sus.ipynb
-    - References: references.md
+    - References: 
+         - API references: references/api.md
 theme: readthedocs
 plugins:
   - mkdocs-jupyter
