### 2022-12-14 #65 @NickleDave
Great, thank you @Robaina for your quick reply and again for your patience.

We haven't run into this situation before where a package was installed through a user's private channel.
We discussed today how to handle that, including maintainers that have more conda experience, and based on that discussion, we will need to require two things for review:
- [ ] making the package available on a community channel such as bioconda or conda-forge. In your case I think bioconda will be the easiest
- [ ] clearly stating the non-Python dependencies in the README, i.e., all the things you install with conda because you can't get them on pip. In your case I think this is HMMer and Prodigal. I know you mention `HMMer` first thing in the README; what we would need is an explicit section with a title like "prerequisites" before "installation/set up" where you list all the non-Python requirements. Ideally since you want `Pynteny` to be a re-usable tool others depend on, you should put a lower but not an upper cap on them, e.g. "HMMer >=1.5", instead of pinning to a specific version like in your environment.yml, since pinning + upper caps users that may depend on other versions of the same libraries

As I said, we have not faced this situation before, so we do not have a process and documentation in place for it. We are correcting that in our docs now. I totally understand if you are frustrated and you feel like we keep adding requirements for review. Please know that is not our intention.

Given that:
- please if you agree focus on getting Pynteny submitted to bioconda for now. I'm guessing you're already acquainted with [the workflow](https://bioconda.github.io/contributor/workflow.html) exist. Let's focus on that instead of me adding noise to your development with issues about whatever weird Linux I have on my laptop
- and please do just add a "prerequistes" section where you list non-Python dependencies. Fine to include an aside like "these will be installed by conda if you install from the bioconda channel"

How does that sound? I promise not to add any more requests! And I'm sorry for not doing due diligence in the first place. We do think this is in scope and want to get you a review as soon as we can!
