### 2022-12-15 #65 @Robaina
Hi @NickleDave,

> We haven't run into this situation before where a package was installed through a user's private channel. We discussed today how to handle that, including maintainers that have more conda experience, and based on that discussion, we will need to require two things for review:

> * [ ]  making the package available on a community channel such as bioconda or conda-forge. In your case I think bioconda will be the easiest

No problem. I actually had already prepared and submitted a recipe to bioconda and checked that it passed all build tests there. But then canceled the pull request since at the time I thought it didn't make much sense to submit to bioconda before the review (and all the potential changes to the code).

However, it's fine with me to submit Pynteny to bioconda prior to revision. Will do.

> * [ ]  clearly stating the non-Python dependencies in the README, i.e., all the things you install with conda because you can't get them on pip. In your case I think this is HMMer and Prodigal. I know you mention `HMMer` first thing in the README; what we would need is an explicit section with a title like "prerequisites" before "installation/set up" where you list all the non-Python requirements. Ideally since you want `Pynteny` to be a re-usable tool others depend on, you should put a lower but not an upper cap on them, e.g. "HMMer >=1.5", instead of pinning to a specific version like in your environment.yml, since pinning + upper caps users that may depend on other versions of the same libraries

Alright. Will add this section. In principle, conda / bioconda will take care of installing all non-python packages, but it makes sense to clearly state that Pynteny depends on other, non-python packages. Particularly since the repo shows that the code is mostly python. Will add that.

Also, I'll remove "environment.yml". This file is no longer required since Pynteny is installed through a conda channel, and the conda recipe is now available in the repo.

> * please if you agree focus on getting Pynteny submitted to bioconda for now. I'm guessing you're already acquainted with [the workflow](https://bioconda.github.io/contributor/workflow.html) but you can ask in our Discourse about this. We are adding someone who knows Bioconda to our team so we can get help from them too. Let's focus on that instead of me adding noise to your development with issues about whatever weird Linux I have on my laptop

No problem. I'll open a PR in bioconda today or tomorrow.

> * and please do just add a "prerequistes" section where you list non-Python dependencies. Fine to include an aside like "these will be installed by conda if you install from the bioconda channel"

Will do

> How does that sound? I promise not to add any more requests! And I'm sorry for not doing due diligence in the first place. We do think this is in scope and want to get you a review as soon as we can!

No problem. I understand that these requirements are meant to improve the software / review process. Also, I know that there are a lot of changes going on at PyOpenSci right now. No problem!
