+++ environment.yml
@@ -1,19 +1,0 @@
-name: pynteny
-channels:
-  - conda-forge
-  - bioconda
-  - defaults
-  - semidan.robaina
-dependencies:
-  - hmmer=3.3.2
-  - python=3.10
-  - prodigal=2.6.3
-  - pyfastx=0.8.4
-  - biopython=1.79
-  - numpy=1.23.1
-  - pandas=1.4.4
-  - psutil=5.9.2
-  - setuptools=62.3.4
-  - python-wget=3.2
-  - streamlit=1.12.2
-  - streamlit-aggrid=0.3.3
