+++ pyproject_no.toml
@@ -1,38 +1,0 @@
-[build-system]
-requires = ["setuptools", "setuptools-scm"]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "pynteny"
-description = "Synteny-aware hmm searches made easy in Python"
-authors = [
-    {name = "Semidán Robaina Estévez", email = "srobaina@ull.edu.es"},
-]
-maintainers = [
-    {name = "Semidán Robaina Estévez", email = "srobaina@ull.edu.es"},
-]
-readme = "README.md"
-requires-python = ">=3.7"
-keywords = ["synteny", "hmm", "metagenomics", "bioinformatics", "hmmer"]
-license = {text = "Apache-2.0 license"}
-classifiers = [
-    "License :: OSI Approved :: Apache Software License",
-    "Programming Language :: Python :: 3",
-    "Topic :: Scientific/Engineering :: Bio-Informatics",
-    "Natural Language :: English"
-]
-version = "0.0.4"
-
-[project.urls]
-homepage = "http://github.com/robaina/Pynteny"
-documentation = "http://github.com/robaina/Pynteny/wiki"
-repository = "http://github.com/robaina/Pynteny"
-
-[project.scripts]
-pynteny = "pynteny.cli:main"
-
-[tool.setuptools]
-include-package-data = true
-
-[tool.setuptools.packages.find]
-where = ["pynteny"]
