### 2022-12-15 a176d34
added back setup.py for conda-build
