+++ MANIFEST.in
@@ -1,0 +1,2 @@
+include pynteny/app/assets/*
+include pynteny/app/assets/img/*
