+++ setup.py
@@ -14,7 +14,7 @@ MAINTAINER = "Semidán Robaina Estévez"
 MAINTAINER_EMAIL = "srobaina@gmail.com"
 DOWNLOAD_URL = 'http://github.com/robaina/Pynteny'
 LICENSE = 'Creative Commons Attribution 4.0 International'
-VERSION = '0.0.4'
+VERSION = '0.0.5'
 
 setup(name=NAME,
       version=VERSION,
