### 2022-12-15 #65 @NickleDave
> I understand that these requirements are meant to improve the software / review process. Also, I know that there are a lot of changes going on at PyOpenSci right now. No problem!

🙌 🙌 🙌

Thank you so much. We really appreciate your understanding. This has been a bit bumpy but you are really helping us improve our process.

> However, for developers / reviewers I think it's best to add a section in README.md containing info for developers.

Yes, this would be perfect. As long as you list them, good enough, but ideally I can clone the repo, make a conda env with those other dependencies, and then `pip install -e .` inside the root of the repo and have everyhing work. So, yes, basically like a developer set up.

> No problem. I'll open a PR in bioconda today or tomorrow.

Excellent!
Please do go ahead and make a new issue that's a full submission after `Pynteny` is on bioconda and you've updated the readme.
Once you make that submission, I will close this issue.
