+++ conda/meta.yaml
@@ -22,7 +22,7 @@ requirements:
 
   host:
     - python
-    - setuptools
+    - poetry
 
   run:
     - python
