+++ conda/build.sh
@@ -1,1 +1,0 @@
-$PYTHON poetry build
