+++ conda/meta.yaml
@@ -6,24 +6,21 @@ package:
   version: "{{ version }}"
 
 source:
-  git_url: "https://github.com/Robaina/Pynteny.git"
+  url: https://github.com/Robaina/pynteny/archive/v{{ version }}.tar.gz
+  sha256: 2cbc392805d835a53dae5847f82c28c53ff8284605783975ffd073c84d949532
 
 build:
   number: 11
-  
   entry_points:
     - pynteny = pynteny.cli:main
+  script: {{ PYTHON }} -m pip install --no-deps --ignore-installed .
 
 requirements:
-
   build:
     - {{ compiler('c') }}
-    - python =3.10
-
   host:
     - python =3.10
-    - poetry
-
+    - pip
   run:
     - python =3.10
     - hmmer
@@ -46,10 +43,12 @@ test:
 
 about:
   home: "http://github.com/robaina/Pynteny"
-  license: Apache-2.0
+  license: Apache-2.0 license
   summary: "Multiple HMM - search via synteny structures in Python"
   doc_url: "https://robaina.github.io/Pynteny/"
 
 extra:
   recipe-maintainers:
     - Robaina
+  identifiers:
+    - DOI:10.5281/zenodo.7082448
