+++ pyproject.toml
@@ -1,30 +1,0 @@
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
-
-[tool.poetry]
-name = "pynteny"
-version = "0.0.5"
-description = "Synteny-aware hmm searches made easy in Python"
-license = "Apache-2.0"
-authors = ["Semidán Robaina Estévez <srobaina@ull.edu.es>"]
-maintainers = ["Semidán Robaina Estévez <srobaina@ull.edu.es>"]
-readme = "README.md"
-homepage = "http://github.com/robaina/Pynteny"
-repository = "http://github.com/robaina/Pynteny"
-documentation = "https://robaina.github.io/Pynteny"
-keywords = ["synteny", "hmm", "metagenomics", "bioinformatics", "hmmer"]
-classifiers = [
-    "License :: OSI Approved :: Apache Software License",
-    "Programming Language :: Python :: 3",
-    "Topic :: Scientific/Engineering :: Bio-Informatics",
-    "Natural Language :: English"
-]
-include = [
-    "pynteny/app/assets/*",
-    "pynteny/app/assets/img/*"
-]
-[tool.poetry.dependencies]
-python = "^3.7"
-[tool.poetry.scripts]
-pynteny = "pynteny.cli:main"
