+++ conda/meta.yaml
@@ -19,6 +19,7 @@ requirements:
     - {{ compiler('c') }}
   host:
     - python =3.10
+    - poetry
     - pip
   run:
     - python =3.10
