+++ .github/workflows/continuous_integration.yml
@@ -25,8 +25,8 @@ jobs:
         with:
           fetch-depth: 0
       - uses: actions/setup-python@v2
-      - run: pip install --upgrade pip && conda install mamba -n base -c conda-forge
-      - run: mamba env create -n "pynteny-dev" -f "environment.yml" && conda activate pynteny-dev && poetry build && pip install dist/pynteny*.whl
+      - run: pip install --upgrade pip
+      - run: conda env create -n "pynteny-dev" -f "environment.yml" && conda activate pynteny-dev && poetry build && pip install dist/pynteny*.whl
       - run: git config user.name 'github-actions[bot]' && git config user.email 'github-actions[bot]@users.noreply.github.com'
       - name: Run tests
         run: python -m unittest discover tests
