+++ .github/workflows/tests-pynteny.yml
@@ -1,0 +1,41 @@
+name: tests-pynteny
+on:
+  push:
+    branches: [ main ]
+    paths:
+      - 'envs/pynteny-dev.yml'
+      - '.github/workflows/tests-pynteny.yml'
+
+jobs:
+
+  create-env:
+    name: ${{ matrix.os }} 
+    runs-on: ${{ matrix.os }}
+    defaults:
+      run:
+        shell: bash -l {0}
+
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ubuntu-latest, macos-latest]
+
+    steps:
+      - name: checkout repository
+        uses: actions/checkout@v2
+        with:
+          fetch-depth: 0
+
+      - name: create environment with mamba
+        uses: conda-incubator/setup-miniconda@v2
+        with:
+          mamba-version: "*"
+          channels: conda-forge,bioconda,defaults
+          auto-activate-base: false
+          activate-environment: tests_pynteny
+          environment-file: envs/pynteny-dev.yml
+
+      - name: Build & Install Pynteny
+        run: poetry build && pip install dist/pynteny*.whl
+      - name: Run tests
+        run: python -m unittest discover tests
