+++ .github/workflows/continuous_integration.yml
@@ -1,37 +1,0 @@
-# This is a basic workflow to help you get started with Actions
-
-name: CI
-
-# Controls when the workflow will run
-on:
-  # Triggers the workflow on push or pull request events but only for the main branch
-  push:
-    branches: [ main ]
-
-  # Allows you to run this workflow manually from the Actions tab
-  workflow_dispatch:
-
-# A workflow run is made up of one or more jobs that can run sequentially or in parallel
-jobs:
-  # This workflow contains a single job called "build"
-  build:
-    # The type of runner that the job will run on
-    runs-on: ubuntu-latest
-
-    # Steps represent a sequence of tasks that will be executed as part of the job
-    steps:
-      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
-      - uses: actions/checkout@v2
-        with:
-          fetch-depth: 0
-      - uses: actions/setup-python@v2
-      - run: pip install --upgrade pip
-      - run: git config user.name 'github-actions[bot]' && git config user.email 'github-actions[bot]@users.noreply.github.com'
-      - name: Install mamba
-        run: curl micro.mamba.pm/install.sh | bash
-      - name: Set environment
-        run: conda deactivate && micromamba env create -n "pynteny-dev" -f "environment.yml" && conda activate pynteny-dev
-      - name: Build & Install Pynteny
-        run: poetry build && pip install dist/pynteny*.whl
-      - name: Run tests
-        run: python -m unittest discover tests
