### 2022-12-19 523fefd
added CI tests
