+++ .github/workflows/draft-pdf.yml
@@ -1,3 +1,4 @@
+name: JOSS-paper
 on: [push]
 
 jobs:
