+++ .github/workflows/joss.yml
@@ -7,7 +7,7 @@ jobs:
     name: Paper Draft
     steps:
       - name: Checkout
-        uses: actions/checkout@v2
+        uses: actions/checkout@v3
       - name: Open Journals PDF Generator
         uses: openjournals/openjournals-draft-action@master
         with:
