+++ .github/workflows/tests.yml
@@ -43,5 +43,10 @@ jobs:
 
       - name: Build & Install Pynteny
         run: poetry build && pip install dist/pynteny*.whl
-      - name: Run tests
-        run: pynteny --help && python -m unittest discover tests
+
+      - name: Run tests and collect coverage
+        run: pynteny --help
+        run: coverage run --omit pynteny/wrappers.py,pynteny/cli.py,pynteny/subcommands.py,pynteny/utils.py -m unittest discover tests && coverage xml
+
+      - name: Upload  coverage to Codecov
+        uses: codecov/codecov-action@v3
