+++ envs/pynteny-dev.yml
@@ -7,6 +7,7 @@ channels:
 dependencies:
   - python >= 3.10
   - poetry >= 1.3
+  - coverage
   - pip >= 22.3
   - hmmer >= 3.3
   - prodigal >= 2.6
